enhance: update workflow step UX - adding a step auto-adds a generic step. No longer need to specify step type when adding - step types can be updated via dropdown on the step list item
