+++ ui/admin/app/lib/model/workflows.ts
@@ -6,6 +6,14 @@ export type WorkflowBase = AgentBase & {
     output: string;
 };
 
+export const StepType = {
+    Command: "command",
+    If: "if",
+    While: "while",
+    // Template: "template",
+} as const;
+export type StepType = (typeof StepType)[keyof typeof StepType];
+
 export type Step = {
     id: string;
     name: string;
@@ -50,14 +58,24 @@ export type Workflow = EntityMeta &
 export type CreateWorkflow = Partial<WorkflowBase> & Pick<WorkflowBase, "name">;
 export type UpdateWorkflow = WorkflowBase;
 
-export const getDefaultStep = (): Step => ({
-    id: crypto.randomUUID(),
-    name: "",
-    description: "",
-    step: "",
-    cache: false,
-    temperature: 0,
-    tools: [],
-    agents: [],
-    workflows: [],
-});
+export const getDefaultStep = (type: StepType = StepType.Command): Step => {
+    const newStep: Step = {
+        id: crypto.randomUUID(),
+        name: "",
+        description: "",
+        step: "",
+        cache: false,
+        temperature: 0,
+        tools: [],
+        agents: [],
+        workflows: [],
+    };
+
+    if (type === StepType.If) {
+        newStep.if = { condition: "", steps: [], else: [] };
+    } else if (type === StepType.While) {
+        newStep.while = { condition: "", maxLoops: 0, steps: [] };
+    }
+
+    return newStep;
+};
